+++ backendAPI/app/Http/Controllers/CommentTypesController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\CommentTypes;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class CommentTypesController extends Controller
 {
@@ -14,12 +15,20 @@ class CommentTypesController extends Controller
      */
     public function index()
     {
-        try {
-            $commentTypes = CommentTypes::all();
-            return response()->json($commentTypes, 200);
-        } catch (Exception $exception) {
-            return response()->json(['error' => $exception], 500);
+        if (Auth::guard('api')->check()) { // Check if user is logged in
+
+            try {
+                $commentTypes = CommentTypes::all();
+                return response()->json($commentTypes, 200);
+            } catch (Exception $exception) {
+                return response()->json(['error' => $exception], 500);
+            }
+
         }
+        else {
+            return response()->json("Not logged in", 401);
+        }
+
     }
 
     /**
@@ -40,12 +49,20 @@ class CommentTypesController extends Controller
      */
     public function store(Request $request)
     {
-        try {
-            $commentType = CommentTypes::create($request->all());
-            return response()->json($commentType, 201);
-        } catch (Exception $exception) {
-            return response()->json(['error' => $exception], 500);
+        if (Auth::guard('api')->check()) { // Check if user is logged in
+
+            try {
+                $commentType = CommentTypes::create($request->all());
+                return response()->json($commentType, 201);
+            } catch (Exception $exception) {
+                return response()->json(['error' => $exception], 500);
+            }
+
         }
+        else {
+            return response()->json("Not logged in", 401);
+        }
+
     }
 
     /**
@@ -56,11 +73,19 @@ class CommentTypesController extends Controller
      */
     public function show(CommentTypes $commentType)
     {
-        try {
-            return response()->json($commentType, 200);
-        } catch (Exception $exception) {
-            return response()->json(['error' => $exception], 500);
+        if (Auth::guard('api')->check()) { // Check if user is logged in
+
+            try {
+                return response()->json($commentType, 200);
+            } catch (Exception $exception) {
+                return response()->json(['error' => $exception], 500);
+            }
+
         }
+        else {
+            return response()->json("Not logged in", 401);
+        }
+
     }
 
     /**
@@ -83,12 +108,20 @@ class CommentTypesController extends Controller
      */
     public function update(Request $request, CommentTypes $commentType)
     {
-        try {
-            $commentType->update($request->all());
-            return response()->json($commentType, 200);
-        } catch (Exception $exception) {
-            return response()->json(['error' => $exception], 500);
+        if (Auth::guard('api')->check()) { // Check if user is logged in
+
+            try {
+                $commentType->update($request->all());
+                return response()->json($commentType, 200);
+            } catch (Exception $exception) {
+                return response()->json(['error' => $exception], 500);
+            }
+
         }
+        else {
+            return response()->json("Not logged in", 401);
+        }
+
     }
 
     /**
@@ -99,11 +132,19 @@ class CommentTypesController extends Controller
      */
     public function destroy(CommentTypes $commentType)
     {
-        try {
-            $commentType->delete();
-            return response()->json(['message' => 'Deleted'], 205);
-        } catch (Exception $exception) {
-            return response()->json(['error' => $exception], 500);
+        if (Auth::guard('api')->check()) { // Check if user is logged in
+
+            try {
+                $commentType->delete();
+                return response()->json(['message' => 'Deleted'], 205);
+            } catch (Exception $exception) {
+                return response()->json(['error' => $exception], 500);
+            }
+
         }
+        else {
+            return response()->json("Not logged in", 401);
+        }
+
     }
 }
